Extract colour_scale callbacks
